+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/PerformanceConfiguration.kt
@@ -59,6 +59,14 @@ public class PerformanceConfiguration private constructor(public val context: Co
 
     public var tracePropagationUrls: Collection<Pattern> = HashSet()
 
+    public fun addOnSpanEndCallback(spanEndCallback: SpanEndCallback) {
+        spanEndCallbacks.add(spanEndCallback)
+    }
+
+    public fun removeOnSpanEndCallback(spanEndCallback: SpanEndCallback) {
+        spanEndCallbacks.remove(spanEndCallback)
+    }
+
     override fun toString(): String =
         "PerformanceConfiguration(" +
                 "context=$context, " +
@@ -91,6 +99,8 @@ public class PerformanceConfiguration private constructor(public val context: Co
         private const val ENABLED_RELEASE_STAGES = "$BUGSNAG_PERF_NS.ENABLED_RELEASE_STAGES"
         private const val VERSION_CODE_KEY = "$BUGSNAG_PERF_NS.VERSION_CODE"
         private const val APP_VERSION_KEY = "$BUGSNAG_PERF_NS.APP_VERSION"
+        private const val TRACE_PROPAGATION_URLS_KEY = "$BUGSNAG_PERF_NS.TRACE_PROPAGATION_URLS"
+        private const val SERVICE_NAME_KEY = "$BUGSNAG_PERF_NS.SERVICE_NAME"
 
         // Bugsnag Notifier keys that we can read
         private const val BSG_API_KEY = "$BUGSNAG_NS.API_KEY"
@@ -122,43 +132,54 @@ public class PerformanceConfiguration private constructor(public val context: Co
             data: Bundle?,
             apiKeyOverride: String?,
         ): PerformanceConfiguration {
-            return PerformanceConfiguration(ctx).apply {
-                (apiKeyOverride ?: data?.getString(API_KEY, data.getString(BSG_API_KEY)))
-                    ?.also { apiKey = it }
+            val config = PerformanceConfiguration(ctx)
 
-                data?.getString(ENDPOINT_KEY)
-                    ?.also { endpoint = it }
-                data?.getBoolean(AUTO_INSTRUMENT_APP_STARTS_KEY, autoInstrumentAppStarts)
-                    ?.also { autoInstrumentAppStarts = it }
-                data?.getString(AUTO_INSTRUMENT_ACTIVITIES_KEY)
-                    ?.also { autoInstrumentActivities = AutoInstrument.valueOf(it) }
+            (apiKeyOverride ?: data?.getString(API_KEY, data.getString(BSG_API_KEY)))
+                ?.also { config.apiKey = it }
+
+            if (data != null) {
+                data.getString(ENDPOINT_KEY)
+                    ?.also { config.endpoint = it }
+
+                config.autoInstrumentAppStarts = data.getBoolean(
+                    AUTO_INSTRUMENT_APP_STARTS_KEY,
+                    config.autoInstrumentAppStarts,
+                )
+
+                data.getString(AUTO_INSTRUMENT_ACTIVITIES_KEY)
+                    ?.also { config.autoInstrumentActivities = AutoInstrument.valueOf(it) }
 
                 // releaseStage / enabledReleaseStage
-                data?.getString(RELEASE_STAGE_KEY, data.getString(BSG_RELEASE_STAGE_KEY))
-                    ?.also { releaseStage = it }
-                data?.getString(ENABLED_RELEASE_STAGES, data.getString(BSG_ENABLED_RELEASE_STAGES))
-                    ?.also { enabledReleaseStages = it.splitToSequence(',').toSet() }
+                data.getString(RELEASE_STAGE_KEY, data.getString(BSG_RELEASE_STAGE_KEY))
+                    ?.also { config.releaseStage = it }
+                data.getString(ENABLED_RELEASE_STAGES, data.getString(BSG_ENABLED_RELEASE_STAGES))
+                    ?.also { config.enabledReleaseStages = it.splitToSequence(',').toSet() }
 
-                if (data?.containsKey(VERSION_CODE_KEY) == true) {
-                    versionCode = data.getInt(VERSION_CODE_KEY).toLong()
-                } else if (data?.containsKey(BSG_VERSION_CODE_KEY) == true) {
-                    versionCode = data.getInt(BSG_VERSION_CODE_KEY).toLong()
+                if (data.containsKey(VERSION_CODE_KEY)) {
+                    config.versionCode = data.getInt(VERSION_CODE_KEY).toLong()
+                } else if (data.containsKey(BSG_VERSION_CODE_KEY)) {
+                    config.versionCode = data.getInt(BSG_VERSION_CODE_KEY).toLong()
                 }
 
-                if (data?.containsKey(APP_VERSION_KEY) == true) {
-                    appVersion = data.getString(APP_VERSION_KEY)
-                } else if (data?.containsKey(BSG_APP_VERSION_KEY) == true) {
-                    appVersion = data.getString(BSG_APP_VERSION_KEY)
+                if (data.containsKey(APP_VERSION_KEY)) {
+                    config.appVersion = data.getString(APP_VERSION_KEY)
+                } else if (data.containsKey(BSG_APP_VERSION_KEY)) {
+                    config.appVersion = data.getString(BSG_APP_VERSION_KEY)
                 }
+
+                if (data.containsKey(TRACE_PROPAGATION_URLS_KEY)) {
+                    config.tracePropagationUrls = data.getString(TRACE_PROPAGATION_URLS_KEY)
+                        ?.splitToSequence(',')
+                        ?.map { it.toPattern() }
+                        ?.toList()
+                        .orEmpty()
+                }
+
+                data.getString(SERVICE_NAME_KEY)
+                    ?.also { config.serviceName = it }
             }
+
+            return config
         }
-    }
-
-    public fun addOnSpanEndCallback(spanEndCallback: SpanEndCallback) {
-        spanEndCallbacks.add(spanEndCallback)
-    }
-
-    public fun removeOnSpanEndCallback(spanEndCallback: SpanEndCallback) {
-        spanEndCallbacks.remove(spanEndCallback)
     }
 }